Crate anchor_attribute_error

Macros

error

Generates an Error::AnchorError that includes file and line information.

Attribute Macros

error_code

Generates Error and type Result<T> = Result<T, Error> types to be used as return types from Anchor instruction handlers. Importantly, the attribute implements From on the ErrorCode to support converting from the user defined error enum into the generated Error.